Rename 'commands' to 'scenes' and remove unused files
