+++ src/index.ts
@@ -14,13 +14,13 @@ import {
     generateExistingProjectsScene,
     generateGroupingsScene,
     resetInteractionsScene,
-} from './commands';
+} from './scenes';
 import { VercelRequest, VercelResponse } from '@vercel/node';
 import { development, production } from './core';
 import { connectToDatabase } from './db/functions';
 import {} from 'dotenv/config';
-import { addPeopleScene } from './commands/editProject/addPeopleScene';
-import { deletePeopleScene } from './commands/editProject/deletePeopleScene';
+import { addPeopleScene } from './scenes/editProject/addPeopleScene';
+import { deletePeopleScene } from './scenes/editProject/deletePeopleScene';
 
 const BOT_TOKEN = process.env.BOT_TOKEN || '';
 const ENVIRONMENT = process.env.NODE_ENV || '';
