+++ src/scenes/deleteProjectScene.ts
