+++ src/scenes/editProject/addPeopleScene.ts
