+++ src/scenes/editProject/deletePeopleScene.ts
