+++ src/scenes/generateGroupingsScene.ts
