+++ src/scenes/resetInteractionsScene.ts
